+++ cli-tests/tests/missing-lib.test.ts
@@ -3,20 +3,18 @@ import { paths } from '../src/entities';
 
 describe("Set of --detect-missing-libraries tests", () => {
     const zksolcCommand = 'zksolc';
-    const solcCommand = 'solc';
-
 
     //id1753
     describe(`Run ${zksolcCommand} with Sol + --detect-missing-libraries}`, () => {
         const args = [`${paths.pathToBasicSolContract}`, ` --detect-missing-libraries`];
         const result = executeCommand(zksolcCommand, args);
 
-        it("Valid command exit code = 0", () => {
-            expect(result.exitCode).toBe(0);
+        it("Valid command exit code = 1", () => {
+            expect(result.exitCode).toBe(1);
         });
 
         it("Compilation info is presented", () => {
-            expect(result.output).toMatch(/(Compiler run successful)/i);
+            expect(result.output).toMatch(/(Missing deployable libraries detection mode is only supported in standard JSON mode.)/i);
         });
     });
 
@@ -30,7 +28,7 @@ describe("Set of --detect-missing-libraries tests", () => {
         });
 
         it("Compilation info is presented", () => {
-            expect(result.output).toMatch(/(No input sources specified)/i);
+            expect(result.output).toMatch(/(Missing deployable libraries detection mode is only supported in standard JSON mode.)/i);
         });
     });