+++ src/project/mod.rs
@@ -338,7 +338,7 @@ impl Project {
     pub fn compile_to_eravm(
         self,
         optimizer_settings: era_compiler_llvm_context::OptimizerSettings,
-        llvm_options: &[&str],
+        llvm_options: Vec<String>,
         enable_eravm_extensions: bool,
         include_metadata_hash: bool,
         bytecode_encoding: zkevm_assembly::RunningVmEncodingMode,
@@ -356,10 +356,7 @@ impl Project {
                         include_metadata_hash,
                         bytecode_encoding == zkevm_assembly::RunningVmEncodingMode::Testing,
                         optimizer_settings.clone(),
-                        llvm_options
-                            .iter()
-                            .map(|option| (*option).to_owned())
-                            .collect(),
+                        llvm_options.clone(),
                         debug_config.clone(),
                     ),
                     era_compiler_llvm_context::Target::EraVM,
@@ -432,7 +429,7 @@ impl Project {
     pub fn compile_to_evm(
         self,
         optimizer_settings: era_compiler_llvm_context::OptimizerSettings,
-        llvm_options: &[&str],
+        llvm_options: Vec<String>,
         include_metadata_hash: bool,
         debug_config: Option<era_compiler_llvm_context::DebugConfig>,
     ) -> anyhow::Result<EVMBuild> {
@@ -446,10 +443,7 @@ impl Project {
                         Cow::Borrowed(&self),
                         include_metadata_hash,
                         optimizer_settings.clone(),
-                        llvm_options
-                            .iter()
-                            .map(|option| (*option).to_owned())
-                            .collect(),
+                        llvm_options.clone(),
                         debug_config.clone(),
                     ),
                     era_compiler_llvm_context::Target::EVM,