+++ src/tests/mod.rs
@@ -63,6 +63,7 @@ pub fn build_solidity(
         false,
         true,
         false,
+        vec![],
         None,
     )?;
 
@@ -80,7 +81,7 @@ pub fn build_solidity(
 
     let build = project.compile_to_eravm(
         optimizer_settings,
-        &[],
+        vec![],
         false,
         false,
         zkevm_assembly::RunningVmEncodingMode::Production,
@@ -128,6 +129,7 @@ pub fn build_solidity_and_detect_missing_libraries(
         false,
         false,
         false,
+        vec![],
         None,
     )?;
 
@@ -171,7 +173,7 @@ pub fn build_yul(sources: BTreeMap<String, String>) -> anyhow::Result<SolcStanda
     let project = Project::try_from_yul_sources(sources, BTreeMap::new(), None, None)?;
     let build = project.compile_to_eravm(
         optimizer_settings,
-        &[],
+        vec![],
         false,
         false,
         zkevm_assembly::RunningVmEncodingMode::Production,
@@ -214,7 +216,7 @@ pub fn build_yul_standard_json(
     let project = Project::try_from_yul_sources(sources, BTreeMap::new(), solc_version, None)?;
     let build = project.compile_to_eravm(
         optimizer_settings,
-        &[],
+        vec![],
         solc_compiler.is_none(),
         false,
         zkevm_assembly::RunningVmEncodingMode::Production,
@@ -248,7 +250,7 @@ pub fn build_llvm_ir_standard_json(
     let project = Project::try_from_llvm_ir_sources(sources)?;
     let build = project.compile_to_eravm(
         optimizer_settings,
-        &[],
+        vec![],
         true,
         false,
         zkevm_assembly::RunningVmEncodingMode::Production,
@@ -282,7 +284,7 @@ pub fn build_eravm_assembly_standard_json(
     let project = Project::try_from_eravm_assembly_sources(sources)?;
     let build = project.compile_to_eravm(
         optimizer_settings,
-        &[],
+        vec![],
         true,
         false,
         zkevm_assembly::RunningVmEncodingMode::Production,
@@ -326,6 +328,7 @@ pub fn check_solidity_warning(
         false,
         false,
         false,
+        vec![],
         suppressed_warnings,
     )?;